Initial commit in new repository. Version 0.1.0.
[Known Bugs]
Most unit tests are passing, but there are currently some bugs with
LVAL streams that pop up occasionally (Memo and OLE types in Access).

[Licensing]
There was a previous repository that erroneously listed the license for
this as the MPL. Since this is a derivative work off MdbTools, this
repository has been updated to reference the LGPL version 2. If you
have received a previous copy of this library, know that the MPL
licensing was erroneous and the terms of the LGPL apply.